// src/app/modules/cadastros/banco/enums/cadastros-banco-enums.model.ts
/**********************************************************************************************
Code generated with MKL Plug-in version: 40.2.5
Code generated at time stamp: 2019-12-31T10:27:34.608
Copyright: Kerubin - logokoch@gmail.com

WARNING: DO NOT CHANGE THIS CODE BECAUSE THE CHANGES WILL BE LOST IN THE NEXT CODE GENERATION.
***********************************************************************************************/

import { SelectItem } from 'primeng/api';

export interface TipoReceitaDespesa extends SelectItem {
	
  label?: string;
  value: any;
  
}

export interface TipoPlanoContaFinanceiro extends SelectItem {
	
  label?: string;
  value: any;
  
}

export interface SituacaoConciliacaoTrn extends SelectItem {
	
  label?: string;
  value: any;
  
}

export interface SituacaoConciliacao extends SelectItem {
	
  label?: string;
  value: any;
  
}

export interface TipoTransacao extends SelectItem {
	
  label?: string;
  value: any;
  
}

export interface TipoContaBancaria extends SelectItem {
	
  label?: string;
  value: any;
  
}